@@ -1,14 +1,14 @@
 module github.com/bartekpacia/database-tools
 
-go 1.15
+go 1.16
 
 require (
-	cloud.google.com/go/firestore v1.4.0
-	cloud.google.com/go/storage v1.12.0
-	firebase.google.com/go/v4 v4.1.0
+	cloud.google.com/go/firestore v1.5.0
+	cloud.google.com/go/storage v1.14.0
+	firebase.google.com/go/v4 v4.2.0
 	github.com/bbrks/go-blurhash v1.1.1
 	github.com/pkg/errors v0.9.1
 	golang.org/x/image v0.0.0-20201208152932-35266b937fa6
-	google.golang.org/api v0.37.0
+	google.golang.org/api v0.40.0
 	github.com/jdeng/goheif v0.0.0-20200323230657-a0d6a8b3e68f
 )
