@@ -2,11 +2,11 @@ package readers
 
 import (
 	"bufio"
-	"github.com/pkg/errors"
 	"io"
-	"io/ioutil"
 	"os"
 	"strings"
+
+	"github.com/pkg/errors"
 )
 
 // ReadFromFile opens and reads from file at filepath. It gracefully
@@ -18,7 +18,7 @@ func ReadFromFile(filepath string) ([]byte, error) {
 	}
 	defer file.Close()
 
-	fileContent, err := ioutil.ReadAll(file)
+	fileContent, err := io.ReadAll(file)
 	if err != nil {
 		return nil, errors.Errorf("failed to read contents from file %s", filepath)
 	}
@@ -27,14 +27,14 @@ func ReadFromFile(filepath string) ([]byte, error) {
 }
 
 // ReadTextualData reads header and content from file.
-func ReadTextualData(file *os.File) (header string, content string, err error) {
+func ReadTextualData(file io.Reader, filename string) (header string, content string, err error) {
 	reader := bufio.NewReader(file)
 	header, err = reader.ReadString('\n')
 	if err != nil {
 		if errors.Is(err, io.EOF) {
 			err = nil
 		} else {
-			err = errors.Errorf("failed to read header (line 1) from file %s: %v", file.Name(), err)
+			err = errors.Errorf("failed to read header (line 1) from file %s: %v", filename, err)
 			return
 		}
 	}
@@ -45,7 +45,7 @@ func ReadTextualData(file *os.File) (header string, content string, err error) {
 		if errors.Is(err, io.EOF) {
 			err = nil
 		} else {
-			err = errors.Errorf("failed to read 3-slash divider (line 2) from file %s: %v", file.Name(), err)
+			err = errors.Errorf("failed to read 3-slash divider (line 2) from file %s: %v", filename, err)
 			return
 		}
 	}
@@ -55,10 +55,11 @@ func ReadTextualData(file *os.File) (header string, content string, err error) {
 		if errors.Is(err, io.EOF) {
 			err = nil
 		} else {
-			err = errors.Errorf("failed to read content (line 3) from file %s: %v", file.Name(), err)
+			err = errors.Errorf("failed to read content (line 3) from file %s: %v", filename, err)
 			return
 		}
 	}
+	content = strings.TrimSuffix(content, "\n")
 
 	return
 }
