@@ -2,10 +2,11 @@ package internal
 
 import (
 	"encoding/json"
-	"github.com/bartekpacia/database-tools/readers"
-	"github.com/pkg/errors"
 	"os"
 	"path/filepath"
+
+	"github.com/bartekpacia/database-tools/readers"
+	"github.com/pkg/errors"
 )
 
 // Story represents a longer piece of text about a particular topic.
@@ -43,6 +44,7 @@ func (s *Story) Parse(lang string) error {
 		return errors.Wrap(err, "makeMarkdownPath")
 	}
 
+	s.imagePaths = make([]string, 0)
 	err = s.makeImagePaths(Compressed)
 	if err != nil {
 		return errors.Wrap(err, "makeImagesPath")
@@ -75,6 +77,10 @@ func (s *Story) makeImagePaths(quality Quality) error {
 	}
 
 	// s.Images were set when the story was parsed from its JSON.
+	if s.Images == nil {
+		s.Images = make([]string, 0)
+	}
+
 	for _, image := range s.Images {
 		absPath := filepath.Join(cwd, "images/", qualityDir, "/"+image+".webp")
 		s.imagePaths = append(s.imagePaths, absPath)
